pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dup"
version = "1.0.0"
description = "Drive Upload Program - Manage Google Drive from the command line"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["google-drive", "cli", "upload", "cloud-storage"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Utilities",
    "Topic :: Internet",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
    "google-auth>=2.17.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.80.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pyinstaller>=5.10.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/dup"
Documentation = "https://github.com/yourusername/dup#readme"
Repository = "https://github.com/yourusername/dup.git"
Issues = "https://github.com/yourusername/dup/issues"

[project.scripts]
dup = "dup.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["dup*"]
exclude = ["tests*", "installer*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
